+++ src/services/file.ts
@@ -31,7 +31,7 @@ class FileService {
     
     
     if (file.type === "success") {
-      const fileSize = Math.trunc(file.size / 1000 / 1000);
+      const fileSize = file.size / 1024 / 1024
       const usageSize = this.filesSizeUsed + fileSize;
       const type = MimeTypes.lookup(file.name).split("/")[0] as string;
       if (fileSize > this.sizeLimit || usageSize > this.sizeLimit) {